requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
python-jose==3.3.0
python-dotenv==1.0.0 